[Refactor] dto 파일 정리 && 코드 정리
