+++ app/src/main/java/org/sopt/at/dto/request/RequestSignInDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.signin
+package org.sopt.at.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
