+++ app/src/main/java/org/sopt/at/dto/request/RequestSignUpDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.signup
+package org.sopt.at.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
