+++ app/src/main/java/org/sopt/at/dto/response/ResponseNicknameDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.my
+package org.sopt.at.dto.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
