+++ app/src/main/java/org/sopt/at/signin/SignInViewModel.kt
@@ -14,6 +14,8 @@ import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 import org.sopt.at.ServicePool
 import org.sopt.at.dto.BaseResponseDto
+import org.sopt.at.dto.request.RequestSignInDto
+import org.sopt.at.dto.response.ResponseSignInDto
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
